feat: added a link to wikipedia
I should see an a element with a corresponding id="tribute-link",
which links to an outside site that contains additional information
about the subject of the tribute page.
HINT: You must give your element an attribute of target and set it to _blank
in order for your link to open in a new tab (i.e. target="_blank").